Modified LatinService to query local DB.
